feat: copy global assets
