+++ .gitignore
@@ -1,1 +1,3 @@
 /dist
+/assets
+Dockerfile
